# api-swagger.yaml
openapi: 3.0.0
info:
  description: Ecommerce API
  version: '1.0.0'
  title: Ecommerce API
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'http://localhost:4000'
...
paths:
...
  /roles/all:
    get:
      tags:
        - roles
      summary: Get all role
      operationId: getAllRole
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Get all role success'
                  data:
                    type: object
                    properties:
                      roles:
                        type: array
                        items:
                          $ref: '#/components/schemas/Role'
                      page:
                        type: number
                        example: 1
                      limit:
                        type: number
                        example: 1
                      total_page:
                        type: number
                        example: 1
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
  /roles:
    post:
      tags:
        - roles
      summary: Create role
      operationId: createRole
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
  /roles/{role_id}:
    put:
      tags:
        - roles
      summary: Update role
      operationId: updateRole
      security:
        - BearerAuth: []
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Update role success'
                  data:
                    $ref: '#/components/schemas/Role'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
  /roles/{id}:
    get:
      tags:
        - roles
      summary: Get detail role
      operationId: getDetailRole
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Get detail role success'
                  data:
                    $ref: '#/components/schemas/Role'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
  /roles/delete/{id}:
    delete:
      tags:
        - roles
      summary: Delete role
      operationId: deleteRole
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Delete role success'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
...
  /products:
    post:
      tags:
        - products
      summary: Create product
      operationId: createProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Create product success'
                  data:
                    $ref: '#/components/schemas/Product'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /products/update/{id}:
    put:
      tags:
        - products
      summary: Update product
      operationId: updateProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: MongoId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductBody'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Update product success'
                  data:
                    $ref: '#/components/schemas/Product'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
...
  /products/all:
    get:
      tags:
        - products
      summary: Get all products
      operationId: getAllProducts
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: productType
          in: query
          required: false
          schema:
            type: string
        - name: productLocation
          in: query
          required: false
          schema:
            type: string
        - name: minPrice
          in: query
          required: false
          schema:
            type: number
        - name: maxPrice
          in: query
          required: false
          schema:
            type: number
        - name: minStar
          in: query
          required: false
          schema:
            type: number
        - name: maxStar
          in: query
          required: false
          schema:
            type: number
        - name: order
          in: query
          required: false
          schema:
            type: string
            default: 'created_at desc'
        - name: statusFilter
          in: query
          required: false
          schema:
            type: string
            default: '1'
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Get all products success'
                  data:
                    type: object
                    properties:
                      products:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'
                      page:
                        type: number
                        example: 1
                      limit:
                        type: number
                        example: 10
                      total_page:
                        type: number
                        example: 1
                      total_count:
                        type: number
                        example: 1
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /products/delete/{id}:
    delete:
      tags:
        - products
      summary: Delete product
      operationId: deleteProduct
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: MongoId
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Delete product success'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /products/liked/me:
    get:
      tags:
        - products
      summary: Get all products liked by me
      operationId: getAllProductLiked
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Get all products liked by me success'
                  data:
                    type: object
                    properties:
                      products:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'
                      page:
                        type: number
                        example: 1
                      limit:
                        type: number
                        example: 10
                      total_page:
                        type: number
                        example: 1

        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /products/viewed/me:
    get:
      tags:
        - products
      summary: Get all products viewed by me
      operationId: getAllProductViewed
      security:
        - BearerAuth: []
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Get all products viewed by me success'
                  data:
                    type: object
                    properties:
                      products:
                        type: array
                        items:
                          $ref: '#/components/schemas/Product'
                      page:
                        type: number
                        example: 1
                      limit:
                        type: number
                        example: 10
                      total_page:
                        type: number
                        example: 1

        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /products/like:
    post:
      tags:
        - products
      summary: Like product
      operationId: likeProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: string
                  format: MongoId
                  example: '679d3b5b6d6d6d6d6d6d6d6d'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Like product success'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /products/unlike:
    post:
      tags:
        - products
      summary: Unlike product
      operationId: unlikeProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: string
                  format: MongoId
                  example: '679d3b5b6d6d6d6d6d6d6d6d'
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Unlike product success'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
  /products/delete-many:
    delete:
      tags:
        - products
      summary: Delete many products
      operationId: deleteManyProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productIds:
                  type: array
                  items:
                    type: string
                    format: MongoId
                  example: ['679d3b5b6d6d6d6d6d6d6d6d', '679d3b5b6d6d6d6d6d6d6d6d']
      responses:
        default:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: 'Delete many products success'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'

  # /orders:
components:
  schemas:
    Unauthorized:
      type: object
      properties:
        message:
          type: string
          example: 'Unauthorized'
...
    Role:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
          example: "69218ec683ccdc34f38abfbd"
        user_id:
          type: string
          format: MongoId
          example: "69218ec683ccdc34f38abfbd"
        name:
          type: string
          example: "Admin"
        permissions:
          type: array
          items:
            type: string
            example: "ADMIN.GRANTED"
        created_at:
          type: string
          format: ISO8601
          example: "2025-11-22T10:21:58.648Z"
        updated_at:
          type: string
          format: ISO8601
          example: "2025-11-22T10:45:46.131Z"
    RoleBody:
      type: object
      properties:
        name:
          type: string
          example: "Admin"
        permissions:
          type: array
          items:
            type: string
            example: "ADMIN.GRANTED"
...
    ProductBody:
      type: object
      properties:
        name:
          type: string
          example: "Product 1"
        image:
          type: string
          example: "https://example.com/image.jpg"
        countInStock:
          type: number
          example: 100000
        price:
          type: number
          example: 10
        discountStartDate:
          type: string
          format: ISO8601
          example: "2025-11-22T10:21:58.648Z"
        discountEndDate:
          type: string
          format: ISO8601
          example: "2025-11-22T10:21:58.648Z"
        type_id:
          type: string
          format: MongoId
          example: "69218ec683ccdc34f38abfbd"
        location_id:
          type: string
          format: MongoId
          example: "69218ec683ccdc34f38abfbd"
    Product:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
          example: "69218ec683ccdc34f38abfbd"
        user_id:
          type: string
          format: MongoId
          example: "69218ec683ccdc34f38abfbd"
        name:
          type: string
          example: "Product 1"
        slug:
          type: string
          example: "product-1"
        image:
          type: string
          example: "https://example.com/image.jpg"
        price:
          type: number
          example: 10
        countInStock:
          type: number
          example: 100000
        description:
          type: string
          example: "Product description"
        discount:
          type: number
          example: 10
        discountStartDate:
          type: string
          format: ISO8601
          example: "2025-11-22T10:21:58.648Z"
        discountEndDate:
          type: string
          format: ISO8601
          example: "2025-11-22T10:21:58.648Z"
        rating:
          type: number
          example: 1
        sold:
          type: number
          example: 1
        type_id:
          type: string
          format: MongoId
          example: "69218ec683ccdc34f38abfbd"
        location_id:
          type: string
          format: MongoId
          example: "69218ec683ccdc34f38abfbd"
        likedBy:
          type: array
          items:
            type: string
            format: MongoId
            example: "69218ec683ccdc34f38abfbd"
        totalLikes:
          type: number
          example: 1
        status:
          type: number
          example: 1
        views:
          type: number
          example: 1
        uniqueViews:
          type: array
          items:
            type: string
            format: MongoId
            example: "69218ec683ccdc34f38abfbd"
        created_at:
          type: string
          format: ISO8601
          example: "2025-11-22T10:21:58.648Z"
        updated_at:
          type: string
          format: ISO8601
          example: "2025-11-22T10:45:46.131Z"
    NotFound:
      type: object
      properties:
        message:
          type: string
          example: 'Not found'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT